nit: refactor directories
