+++ packages/indexer/src/services/index.ts
@@ -1,3 +1,3 @@
 export * as bundles from "./bundles";
 export * as spokeProcessor from "./spokePoolProcessor";
-export * as bundleBuilder from "./bundleBuilder";
+export * as bundleBuilder from "./BundleBuilderService";
